Fixed a DeprecationWarning (Django 1.2+).
Since Django 1.2, using short versions of built-in db backends names
in ``settings.py`` is deprecated; the full import path for the backend
should be used, instead.

When running tests, the following message was printed to the console:

``DeprecationWarning: Short names for ENGINE in database
configurations are deprecated. Prepend default.ENGINE with
'django.db.backends.'``

Now, it's gone.